--- client/src/components/Form/FormBtn.js
@@ -1,9 +1,0 @@
-import React from "react";
-
-export function FormBtn(props) {
-  return (
-    <button {...props} style={{ float: "right", marginBottom: 10 }} className="btn btn-success">
-      {props.children}
-    </button>
-  );
-}
